refactor: extract modal to be reusable and update styling for modal features
